Task- 63 Creating a Directory TypeAhead,Option.Svelte moved,Removing Button, and other minor changes
